major refactoring to make code sane to manage. split company stuff into a separate package
