+++ bin/simulate_hiring_risk.py
@@ -8,7 +8,7 @@ import datetime
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
-from a_model.datascope import Datascope
+from a_model.company import Company
 from a_model.argparsers import HiringParser
 from a_model import utils
 from a_model import decorators
@@ -17,8 +17,8 @@ from a_model import decorators
 parser = HiringParser(description=__doc__)
 args = parser.parse_args()
 
-# instantiate datascope
-datascope = Datascope(today=args.today)
+# instantiate company
+company = Company(today=args.today)
 
 
 # simulate finances in our current situation and by adding up to n_n00bs new
@@ -28,22 +28,22 @@ def get_all_n00b_outcomes():
     all_n00b_outcomes = []
     for n00b in range(0, args.n_n00bs+1):
         if n00b > 0:
-            datascope.add_person("n00b_%d" % n00b)
-        monthly_cash_outputs, _, _ = datascope.simulate_monthly_cash(
+            company.add_person("n00b_%d" % n00b)
+        monthly_cash_outputs, _, _ = company.simulate_monthly_cash(
             n_months=args.n_months,
             n_universes=args.n_universes,
             verbose=args.verbose,
         )
 
         # create a data structure to store the information in a relevant way
-        n00b_outcomes = datascope.get_outcomes_in_month(0, [])
+        n00b_outcomes = company.get_outcomes_in_month(0, [])
         all_n00b_outcomes.append(n00b_outcomes)
         for k in n00b_outcomes:
             n00b_outcomes[k] = []
 
         # calculate the outcomes month by month
         for month in range(args.n_months):
-            outcomes = datascope.get_outcomes_in_month(
+            outcomes = company.get_outcomes_in_month(
                 month,
                 monthly_cash_outputs,
             )
@@ -59,7 +59,7 @@ plt.rcParams["figure.figsize"] = [b, a]
 
 # create a figure for each outcome
 figure, axes = plt.subplots(len(all_n00b_outcomes[0]), sharex=True)
-time = [t for t in datascope.iter_future_months(args.n_months)]
+time = [t for t in company.iter_future_months(args.n_months)]
 for n_n00bs in range(len(all_n00b_outcomes)):
 
     # configure a few parameters for this particular number of n00bs
