+++ bin/sync_quickbooks_gdrive.py
@@ -6,17 +6,17 @@ spreadsheet.
 """
 
 from a_model import reports
-from a_model.datascope import Datascope
+from a_model.company import Company
 from a_model.argparsers import SyncParser
 
 # parse command line arguments
 parser = SyncParser(description=__doc__)
 args = parser.parse_args()
 
-# get some credentials from the datascope object
-datascope = Datascope(today=args.today, add_people=False)
-username = datascope.config.get('quickbooks', 'username')
-password = datascope.config.get('quickbooks', 'password')
+# get some credentials from the company object
+company = Company(today=args.today, add_people=False)
+username = company.config.get('quickbooks', 'username')
+password = company.config.get('quickbooks', 'password')
 
 # download and sync the reports
 reports.cache_quickbooks_locally(username, password, reports=args.reports)
